openmesh-collectors/requirements.txt
aiohttp>=3.8.1
aiokafka>=0.7.2
aioredis>=2.0.1
confluent_kafka>=1.8.2
faust-streaming
psutil>=5.9.1
python-dotenv>=0.21.0
python_dateutil>=2.8.2
python_schema_registry_client==2.4.0
requests>=2.22.0
websockets>9.1
yapic.json>=1.7.3
uvloop
web3==6.7.0
cachetools
authlib